Update to SPDX license names
Apply updates from running the new convert-spdx-licenses script
from upstream.  This is not currently a hard requirement from
upstream, but futureproofs for when the license name mapping is
finally removed.  Additionally, a few license declarations of just
"BSD" have been updated with the more specifc "BSD-3-Clause" after
inspecting the licenses in the source, as the non-specific name
now triggers a warning.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ia86d412fe76ff3217456fcea81cb44f68bc1ced8